Update: Make FavoritteButton a component
