+++ components/FavoriteButton/index.module.css
@@ -1,0 +1,21 @@
+.favoritesBtn {
+  border: none;
+  background-color: transparent;
+  margin-left: 25px;
+  vertical-align: -2px;
+}
+
+.favoritesBtn:hover {
+  opacity: 0.7;
+  cursor: pointer;
+}
+
+.favoritesBtnIconOn {
+  color: var(--palevioletred);
+  font-size: 1.2rem;
+}
+
+.favoritesBtnIconOff {
+  color: var(--gray);
+  font-size: 1.2rem;
+}
