+++ components/PlaceInfo/index.module.css
@@ -48,27 +48,6 @@
   vertical-align: -2px;
 }
 
-.favoritesBtn {
-  border: none;
-  background-color: transparent;
-  margin: 0 10px 0 auto;
-}
-
-.favoritesBtn:hover {
-  opacity: 0.7;
-  cursor: pointer;
-}
-
-.favoritesBtnIconOn {
-  color: var(--palevioletred);
-  font-size: 1.2rem;
-}
-
-.favoritesBtnIconOff {
-  color: var(--gray);
-  font-size: 1.2rem;
-}
-
 .footer {
   margin-top: 5px;
   display: flex;
